+++ typescript/aptos-indexer-protos/src/aptos/util/timestamp/timestamp.ts
@@ -71,12 +71,12 @@ export const Timestamp = {
     source: AsyncIterable<Timestamp | Timestamp[]> | Iterable<Timestamp | Timestamp[]>,
   ): AsyncIterable<Uint8Array> {
     for await (const pkt of source) {
-      if (Array.isArray(pkt)) {
-        for (const p of pkt) {
+      if (globalThis.Array.isArray(pkt)) {
+        for (const p of (pkt as any)) {
           yield* [Timestamp.encode(p).finish()];
         }
       } else {
-        yield* [Timestamp.encode(pkt).finish()];
+        yield* [Timestamp.encode(pkt as any).finish()];
       }
     }
   },
@@ -87,12 +87,12 @@ export const Timestamp = {
     source: AsyncIterable<Uint8Array | Uint8Array[]> | Iterable<Uint8Array | Uint8Array[]>,
   ): AsyncIterable<Timestamp> {
     for await (const pkt of source) {
-      if (Array.isArray(pkt)) {
-        for (const p of pkt) {
+      if (globalThis.Array.isArray(pkt)) {
+        for (const p of (pkt as any)) {
           yield* [Timestamp.decode(p)];
         }
       } else {
-        yield* [Timestamp.decode(pkt)];
+        yield* [Timestamp.decode(pkt as any)];
       }
     }
   },