+++ typescript/aptos-indexer-protos/CHANGELOG.md
@@ -4,5 +4,8 @@ All notable changes to the Aptos Indexer Protos will be captured in this file. T
 
 ## Unreleased
 
+## 0.2.3
+- Minor changes due to upgrading to protoc-gen-ts version 0.8.7.
+
 ## 0.2.2
 - Initial release.